+++ lib/util/extensions/enemy/enemy_extensions.dart
@@ -102,11 +102,15 @@ extension EnemyExtensions on Enemy {
   }
 
   /// Checks whether the player is within range. If so, move to it.
+  /// [visionAngle] in radians
+  /// [angle] in radians. is automatically picked up using the component's direction.
   void seeAndMoveToAttackRange({
     required Function(Player) positioned,
     VoidCallback? notObserved,
     VoidCallback? observed,
     double radiusVision = 32,
+    double? visionAngle,
+    double? angle,
     double? minDistanceFromPlayer,
     bool runOnlyVisibleInScreen = true,
   }) {
@@ -114,6 +118,8 @@ extension EnemyExtensions on Enemy {
 
     seePlayer(
       radiusVision: radiusVision,
+      visionAngle: visionAngle,
+      angle: angle,
       observed: (player) {
         observed?.call();
         this.positionsItselfAndKeepDistance(