+++ lib/util/extensions/npc/npc_extensions.dart
@@ -2,6 +2,7 @@ import 'dart:math';
 import 'dart:ui';
 
 import 'package:bonfire/bonfire.dart';
+import 'package:bonfire/geometry/shape.dart';
 
 ///
 /// Created by
@@ -18,37 +19,49 @@ import 'package:bonfire/bonfire.dart';
 extension NpcExtensions on Npc {
   /// This method we notify when detect the player when enter in [radiusVision] configuration
   /// Method that bo used in [update] method.
-  void seePlayer({
+  /// [visionAngle] in radians
+  /// [angle] in radians. is automatically picked up using the component's direction.
+  Shape? seePlayer({
     required Function(Player) observed,
     VoidCallback? notObserved,
     double radiusVision = 32,
+    double? visionAngle,
+    double? angle,
   }) {
     Player? player = gameRef.player;
     if (player == null || player.isDead) {
       notObserved?.call();
-      return;
+      return null;
     }
-    this.seeComponent(
+    return this.seeComponent(
       player,
       observed: (c) => observed(c as Player),
       notObserved: notObserved,
       radiusVision: radiusVision,
+      visionAngle: visionAngle,
+      angle: angle ?? this.lastDirection.toRadians(),
     );
   }
 
   /// Checks whether the player is within range. If so, move to it.
-  void seeAndMoveToPlayer({
+  /// [visionAngle] in radians
+  /// [angle] in radians. is automatically picked up using the component's direction.
+  Shape? seeAndMoveToPlayer({
     required Function(Player) closePlayer,
     VoidCallback? notObserved,
     VoidCallback? observed,
     double radiusVision = 32,
     double margin = 10,
+    double? visionAngle,
+    double? angle,
     bool runOnlyVisibleInScreen = true,
   }) {
-    if (runOnlyVisibleInScreen && !this.isVisible) return;
+    if (runOnlyVisibleInScreen && !this.isVisible) return null;
 
-    seePlayer(
+    return seePlayer(
       radiusVision: radiusVision,
+      visionAngle: visionAngle,
+      angle: angle,
       observed: (player) {
         observed?.call();
         this.followComponent(
@@ -68,11 +81,15 @@ extension NpcExtensions on Npc {
   }
 
   /// Checks whether the player is within range. If so, move to it.
+  /// [visionAngle] in radians
+  /// [angle] in radians. is automatically picked up using the component's direction.
   void seeAndMoveToEnemy({
     required Function(Enemy) closeEnemy,
     VoidCallback? notObserved,
     VoidCallback? observed,
     double radiusVision = 32,
+    double? visionAngle,
+    double? angle,
     double margin = 10,
     bool runOnlyVisibleInScreen = true,
   }) {
@@ -80,6 +97,8 @@ extension NpcExtensions on Npc {
 
     seeComponentType<Enemy>(
       radiusVision: radiusVision,
+      visionAngle: visionAngle,
+      angle: angle ?? lastDirection.toRadians(),
       observed: (enemy) {
         observed?.call();
         this.followComponent(
@@ -99,11 +118,15 @@ extension NpcExtensions on Npc {
   }
 
   /// Checks whether the ally is within range. If so, move to it.
+  /// [visionAngle] in radians
+  /// [angle] in radians. is automatically picked up using the component's direction.
   void seeAndMoveToAlly({
     required Function(Ally) closeAlly,
     VoidCallback? notObserved,
     VoidCallback? observed,
     double radiusVision = 32,
+    double? visionAngle,
+    double? angle,
     double margin = 10,
     bool runOnlyVisibleInScreen = true,
   }) {
@@ -111,6 +134,8 @@ extension NpcExtensions on Npc {
 
     seeComponentType<Ally>(
       radiusVision: radiusVision,
+      visionAngle: visionAngle,
+      angle: angle ?? lastDirection.toRadians(),
       observed: (ally) {
         observed?.call();
         this.followComponent(