+++ lib/util/extensions/player/player_extensions.dart
@@ -6,16 +6,22 @@ import 'package:flutter/widgets.dart';
 extension PlayerExtensions on Player {
   /// This method we notify when detect the enemy when enter in [radiusVision] configuration
   /// Method that bo used in [update] method.
+  /// [visionAngle] in radians
+  /// [angle] in radians. is automatically picked up using the component's direction.
   void seeEnemy({
     required Function(List<Enemy>) observed,
     VoidCallback? notObserved,
     double radiusVision = 32,
+    double? visionAngle,
+    double? angle,
   }) {
     if (isDead) return;
     this.seeComponentType<Enemy>(
       observed: observed,
       notObserved: notObserved,
       radiusVision: radiusVision,
+      angle: angle ?? this.lastDirection.toRadians(),
+      visionAngle: visionAngle,
     );
   }